+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/controller/EnvStatusController.java
@@ -23,6 +23,7 @@ import de.gematik.test.tiger.testenvmgr.data.BannerType;
 import de.gematik.test.tiger.testenvmgr.data.TigerEnvStatusDto;
 import de.gematik.test.tiger.testenvmgr.data.TigerServerStatusDto;
 import de.gematik.test.tiger.testenvmgr.env.*;
+import io.micrometer.common.util.StringUtils;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -109,11 +110,15 @@ public class EnvStatusController implements TigerUpdateListener {
                 if (scenarioUpdate.getStatus() != TestResult.UNUSED) {
                   scenario.setStatus(scenarioUpdate.getStatus());
                 }
-                if (!scenarioUpdate.getDescription().isEmpty()) {
+                if (!StringUtils.isBlank(scenarioUpdate.getDescription())) {
                   scenario.setDescription(scenarioUpdate.getDescription());
                 }
-                scenario.setExampleKeys(scenarioUpdate.getExampleKeys());
-                scenario.setExampleList(scenarioUpdate.getExampleList());
+                if (scenarioUpdate.getExampleKeys() != null) {
+                  scenario.setExampleKeys(scenarioUpdate.getExampleKeys());
+                }
+                if (scenarioUpdate.getExampleList() != null) {
+                  scenario.setExampleList(scenarioUpdate.getExampleList());
+                }
                 scenario.setVariantIndex(scenarioUpdate.getVariantIndex());
                 scenario.setDryRun(scenarioUpdate.isDryRun());
                 fillInStepData(scenarioUpdate, scenario);
@@ -130,11 +135,15 @@ public class EnvStatusController implements TigerUpdateListener {
             (stepKey, stepUpdate) -> {
               if (scenario.getSteps().containsKey(stepKey)) {
                 StepUpdate step = scenario.getSteps().get(stepKey);
-                fillInStatus(scenario, step, stepUpdate);
-                if (!stepUpdate.getDescription().isEmpty()) {
+                if (stepUpdate.getStatus() != null) {
+                  fillInStatus(scenario, step, stepUpdate);
+                }
+                if (!StringUtils.isBlank(stepUpdate.getDescription())) {
                   step.setDescription(stepUpdate.getDescription());
                 }
-                step.setTooltip(stepUpdate.getTooltip());
+                if (!StringUtils.isBlank(stepUpdate.getTooltip())) {
+                  step.setTooltip(stepUpdate.getTooltip());
+                }
                 step.setStepIndex(stepUpdate.getStepIndex());
                 fillInMetaData(step, stepUpdate);
               } else {